requirements.txt
telethon
python-dotenv
fastapi
uvicorn
pandas
sqlalchemy
asyncpg
dbt-core
dbt-postgres
ultralytics     # for YOLOv8
psycopg2-binary
dagster
dagster
dagster-webserver
